Ajout des sessions pour les contrôles d'url

+++ src/main/java/fr/arnaud/ludovic/themes/controllers/commande/CommandeDeleteTheme.java
@@ -2,6 +2,7 @@ package fr.arnaud.ludovic.themes.controllers.commande;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import fr.arnaud.ludovic.themes.modeles.dto.ThemeDTO;
 import fr.arnaud.ludovic.themes.modeles.entities.Theme;
@@ -13,7 +14,13 @@ import fr.arnaud.ludovic.themes.services.impl.ServiceThemeImpl;
  * Classe CommandeDeleteTheme qui implémente Commande
  */
 public class CommandeDeleteTheme implements Commande {
-	
+
+	/**
+	 * Appel de HttpSession pour vérifier que l'utilisateur est loggé, sinon renvoie
+	 * sur la page de login
+	 */
+	HttpSession session;
+
 	/** Implémente le Service de Theme */
 	private ServiceTheme service = new ServiceThemeImpl();
 
@@ -27,18 +34,24 @@ public class CommandeDeleteTheme implements Commande {
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) {
 		String themeId = request.getParameter("themeId");
-		System.out.println(themeId);
-		
-		try {
-			ThemeDTO themeDto = new ThemeDTO(Integer.valueOf(themeId));
-			service.deleteThemeById(themeDto);
-			
-			return "redirect";
-			
-		} catch (Exception e) {
-			request.setAttribute("msg", e.getMessage());
-			return "themes";
+
+		session = request.getSession(true);
+
+		if (session.getAttribute("isConnected").equals(true)) {
+			try {
+				ThemeDTO themeDto = new ThemeDTO(Integer.valueOf(themeId));
+				service.deleteThemeById(themeDto);
+
+				return "redirect";
+
+			} catch (Exception e) {
+				request.setAttribute("msg", e.getMessage());
+				return "themes";
+			}
+		} else {
+			return "login";
 		}
+
 	}
 
 }
